+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 ### 4.1.0 (Next)
+* [#265](https://github.com/alexa-js/alexa-app/pull/265): Added support for Display.ElementSelected - [@jstreb](https://github.com/jstreb)
 * [#97](https://github.com/alexa-js/alexa-app/pull/97): Added support for custom directives - [@ajcrites](https://github.com/ajcrites).
 * [#248](https://github.com/alexa-js/alexa-app/pull/248): Add support for intent and slot `confirmationStatus` - [@sbonami](https://github.com/sbonami).
 * [#250](https://github.com/alexa-js/alexa-app/pull/250): Add support for handling and/or delegating Dialog - [@sbonami](https://github.com/sbonami).